+++ app/__init__.py
@@ -1,7 +1,8 @@
-from flask import Flask
+from flask import Flask, app, redirect, url_for
 from flask_sqlalchemy import SQLAlchemy
 from os import path
 from flask_login import LoginManager
+
 
 db = SQLAlchemy()
 DB_NAME = "database.db"
@@ -11,14 +12,20 @@ def create_app():
     app.config['SECRET_KEY'] = 'secretkey'
     app.config['SQLALCHEMY_DATABASE_URI'] = f'sqlite:///{DB_NAME}'
     db.init_app(app)
+    
+    from app.blueprints.auth.routes import auth as auth_blueprint
+    from app.blueprints.admin.routes import admin as admin_blueprint
+    from app.blueprints.client.routes import client as client_blueprint
 
-    from .views import views
-    from .auth import auth
+    app.register_blueprint(auth_blueprint, url_prefix='/auth')
+    app.register_blueprint(admin_blueprint, url_prefix='/admin')
+    app.register_blueprint(client_blueprint, url_prefix='/client')
+    
+    @app.route('/')
+    def index():
+        return redirect(url_for('auth.home'))
 
-    app.register_blueprint(views, url_prefix='/')
-    app.register_blueprint(auth, url_prefix='/')
-
-    from .models import User, Product, Order
+    from .models import User
 
     create_database(app)
 
@@ -31,6 +38,8 @@ def create_app():
         return User.query.get(int(id))
 
     return app
+
+
 
 def create_database(app):
     if not path.exists('instance/' + DB_NAME):
