+++ app/templates/admin/prod_edit.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Edit Product{% endblock %}
 
@@ -24,7 +24,7 @@
 <hr>
 
 <h3>Delete Product</h3>
-<form action="{{ url_for('auth.delprod', id=product.id) }}" method="POST" onsubmit="return confirm('Are you sure you want to delete this product?');">
+<form action="{{ url_for('admin.delprod', id=product.id) }}" method="POST" onsubmit="return confirm('Are you sure you want to delete this product?');">
     <button type="submit" class="btn btn-danger">Delete Product</button>
 </form>
