+++ app/templates/admin/showdata.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Show Data - SweetOrder{% endblock %}
 
@@ -28,7 +28,7 @@
     <div class="my-5"></div>
     <h2 class="mb-4">Products List</h2>
     {% if products|length == 0 %}
-    <a href="{{ url_for('auth.adicionar_produtos_padrao') }}" class="btn btn-warning">Adicionar Produtos Padrão</a>
+    <a href="{{ url_for('admin.adicionar_produtos_padrao') }}" class="btn btn-warning">Adicionar Produtos Padrão</a>
     {% endif %}
         <table class="table table-striped table-bordered">
         <thead class="thead-dark">
