+++ app/templates/auth/index.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Home - SweetOrder{% endblock %}
