+++ app/templates/auth/login.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Login - SweetOrder{% endblock %}
