+++ app/templates/auth/logout.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Logout - SweetOrder{% endblock %}
