+++ app/templates/auth/sign_in.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Sign-in - SweetOrder{% endblock %}
