+++ app/templates/client/cart.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 {% block content %}
 <div class="container">
     <h2>Your Cart</h2>
@@ -12,7 +12,7 @@
                     <th>Quantity</th>
                     <th>Total</th>
                     <th style="width: 160px; text-align: left;">
-                        <form action="{{ url_for('auth.clear_cart') }}" method="POST">
+                        <form action="{{ url_for('client.clear_cart') }}" method="POST">
                             <button type="submit" class="btn btn-outline-danger btn-sm">Esvaziar Carrinho</button>
                         </form>
                     </th>
@@ -26,14 +26,14 @@
                     </td>
                     <td>{{ item.name }}</td>
                     <td>
-                        <a href="{{ url_for('auth.decrease_quantity', product_id=product_id) }}" class="btn btn-sm btn-outline-warning">-</a>
+                        <a href="{{ url_for('client.decrease_quantity', product_id=product_id) }}" class="btn btn-sm btn-outline-warning">-</a>
                         {{ item.quantity }}
-                        <a href="{{ url_for('auth.increase_quantity', product_id=product_id) }}" class="btn btn-sm btn-outline-primary">+</a>
+                        <a href="{{ url_for('client.increase_quantity', product_id=product_id) }}" class="btn btn-sm btn-outline-primary">+</a>
                     </td>
                     <td>${{ item.price }}</td>
                     <td>${{ item.price * item.quantity }}</td>
                     <td>
-                        <form action="{{ url_for('auth.remove_from_cart', product_id=product_id) }}" method="POST" style="display:inline;">
+                        <form action="{{ url_for('client.remove_from_cart', product_id=product_id) }}" method="POST" style="display:inline;">
                             <button type="submit" class="btn btn-danger btn-sm">Remover</button>
                         </form>
                     </td>
@@ -45,7 +45,7 @@
             <h3 style="margin-right: 20px;">Total: ${{ total }}</h3>
         </div>
         <div class="d-flex justify-content-end">
-            <form action="{{ url_for('auth.checkout') }}" method="POST">
+            <form action="{{ url_for('client.checkout') }}" method="POST">
                 <button type="submit" class="btn btn-success">Checkout</button>
             </form>
         </div>
