+++ app/templates/client/my_orders.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 {% block title %}My Orders{% endblock %}
 {% block content %}
 <div class="container">
@@ -13,7 +13,7 @@
             </div>
             <div class="d-flex align-items-center">
                 {% if order.status == 'Pendente' %}
-                <form action="{{ url_for('auth.cancel_order', order_id=order.id) }}" method="POST">
+                <form action="{{ url_for('client.cancel_order', order_id=order.id) }}" method="POST">
                     <button type="submit" class="btn btn-sm btn-danger">Cancelar Pedido</button>
                 </form>
                 {% endif %}
