+++ app/templates/client/products.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "shared/base.html" %}
 
 {% block title %}Products - SweetOrder{% endblock %}
 
@@ -21,10 +21,10 @@
                     <div class="d-flex justify-content-between align-items-center">
                         <small class="text-muted">${{ product.price }}</small>
                         {% if current_user.admin %}
-                        <a href="{{ url_for('auth.editprod', id=product.id) }}" class="btn btn-primary">Editar</a>
+                        <a href="{{ url_for('admin.editprod', id=product.id) }}" class="btn btn-primary">Editar</a>
                         {% endif %}
                         {% if not current_user.admin %}
-                        <a href="{{ url_for('auth.add_to_cart', product_id=product.id) }}" class="btn btn-primary"><i class="bi bi-cart"></i> Adicionar</a>
+                        <a href="{{ url_for('client.add_to_cart', product_id=product.id) }}" class="btn btn-primary"><i class="bi bi-cart"></i> Adicionar</a>
                         {% endif %}
                     </div>
                 </div>
@@ -49,10 +49,10 @@
                     </div>
                     <div class="modal-footer">
                         {% if not current_user.admin %}
-                        <a href="{{ url_for('auth.add_to_cart', product_id=product.id) }}" class="btn btn-primary">Adicionar</a>
+                        <a href="{{ url_for('client.add_to_cart', product_id=product.id) }}" class="btn btn-primary">Adicionar</a>
                         {% endif %}
                         {% if current_user.admin %}
-                        <a href="{{ url_for('auth.editprod', id=product.id) }}" class="btn btn-primary">Editar</a>
+                        <a href="{{ url_for('admin.editprod', id=product.id) }}" class="btn btn-primary">Editar</a>
                         {% endif %}
                         <button type="button" class="btn btn-secondary" data-dismiss="modal">Fechar</button>
                     </div>
