--- app/templates/my_profile.html
@@ -1,54 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Meus dados - SweetOrder{% endblock %}
-
-{% block content %} <h1>Meus dados </h1> 
-
-<form class="form-horizontal" method="POST" action="{{ url_for('auth.update_profile') }}">
-
-    <div class="form-group">
-        <label for="email">E-mail:</label>
-        <input type="email" class="form-control" id="email" name="email" placeholder="{{current_user.email}}">
-    </div>
-    <div class="form-group">
-        <label for="password">Senha:</label>
-        <input type="password" class="form-control" id="password" name="password" placeholder="Pelo menos  6 caracteres">
-    </div>
-    <div class="form-group">
-        <label for="address">Endereço:</label>
-        <input type="text" class="form-control" id="address" name="address" placeholder="{{current_user.address}}">
-    </div>
-    <div class="form-group">
-        <label for="phone">Telefone:</label>
-        <input type="text" class="form-control" id="phone" name="phone" placeholder="{{current_user.phone}}">
-    </div>
-    <button type="submit" class="btn btn-primary" id="updateBtn" disabled>Atualizar</button>
-</form>
-
-<script>
-    
-    const updateBtn = document.getElementById('updateBtn');
-    const formFields = document.querySelectorAll('#email, #password, #address, #phone');
-    
-    
-    function checkForm() {
-        let hasChange = false;
-        
-        
-        formFields.forEach(field => {
-            if (field.value.trim() !== '' && field.value !== field.defaultValue) {
-                hasChange = true;
-            }
-        });
-
-        
-        updateBtn.disabled = !hasChange;
-    }
-
-    
-    formFields.forEach(field => {
-        field.addEventListener('input', checkForm);
-    });
-</script>
-
-{% endblock %}
